requirements.txt
asn1crypto==0.24.0
Authlib==0.11
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.2
Flask-Login==0.4.1
Flask-MySQLdb==0.2.0
Flask-Session==0.3.1
Flask-OAuth==0.12
Flask-SocketIO==4.2.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
google==2.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
-e git+https://github.com/heroku/kafka-helper@70aa539540e7d6517f0c3c1900617767ab198b6c#egg=kafka_helper
MarkupSafe==1.1.1
numpy==1.14.2
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.0.1
pipenv==2018.11.26
py-dateutil==2.2
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pyOpenSSL==19.0.0
python-engineio==3.9.0
python-socketio==4.3.0
pytz==2017.3
requests==2.21.0
requests-oauthlib==1.2.0
rfc3339==6.0
rsa==4.0
scikit-learn==0.19.1
scipy==1.0.1
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
uritemplate==3.0.0
urllib3==1.24.3
virtualenv==15.1.0
virtualenv-clone==0.5.3
Werkzeug==0.15.5
WTForms==2.2.1
alabaster==0.7.12
Babel==2.7.0
bootstrap-py==0.8.2
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
docutils==0.15.2
Flask-Session==0.3.1
gitdb2==2.0.5
imagesize==1.1.0
infinity==1.4
intervals==0.8.1
intervals==0.8.1
packaging==19.2
pguard==0.4.0
Pygments==2.4.2
pyparsing==2.4.2
WTForms-Components==0.10.4
XlsxWriter==1.2.1
validators==0.14.0
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-applehelp==1.0.1
snowballstemmer==1.9.1
smmap2==2.0.5
